TgYouTube/requirements.txt
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
mutagen==1.46.0
pycryptodomex==3.18.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
pytube==15.0.0
requests==2.31.0
urllib3==2.0.3
websockets==11.0.3
yt-dlp==2023.7.6
